requirements.txt
python-dotenv>=1.0.0
supabase>=2.0.0
logtail-python>=0.3.0
Pillow>=10.0.0
pdf2image>=1.16.0
PyMuPDF>=1.24.0
httpx>=0.27.0

